# Revert "optimize zipper performance"