Running against a large elixir umbrella app, the optimization showed a 50% slow down :O

```
➜ # v1.4.2
➜ hyperfine 'mix format' -i --warmup 1 --runs 32 # v1.4.2
Benchmark 1: mix format
  Time (mean ± σ):      1.094 s ±  0.027 s    [User: 0.781 s, System: 0.818 s]
  Range (min … max):    1.048 s …  1.155 s    32 runs

➜ # styler main
➜ hyperfine 'mix format' -i --warmup 1 --runs 32 
Benchmark 1: mix format
  Time (mean ± σ):      1.534 s ±  0.037 s    [User: 1.009 s, System: 0.979 s]
  Range (min … max):    1.475 s …  1.655 s    32 runs
```