@novaugust 2025-06-12
cc @doorgan - PR'd this so you could see :O this reverts the tuple-instead-of-map optimization in styler's zipper. it seems that Zipper.traverse is a poor benchmark, but honestly the 50% slowdown means something very odd is happening in there. maybe there's a singular issue i could identify that'd speed things back up.

but honestly, `mix format` bottlenecks on IO so hard that running with-or-without styler makes no difference to user-perceived runtime so i might walk from optimizations for now ;)

```
➜  # running with no styler
➜  hyperfine 'mix format' -i --warmup 1 --runs 32 # no styler
Revert "optimize zipper performance"
Benchmark 1: mix format
  Time (mean ± σ):      1.095 s ±  0.041 s    [User: 0.788 s, System: 0.830 s]
  Range (min … max):    1.046 s …  1.255 s    32 runs
```