@doorgan 2025-06-12
Huh, interesting, please do keep me posted, performance is a tricky thing to deal with